Use uint32 as bit buffer.

+++ dbs/eg_coding.py
@@ -5,9 +5,9 @@ Exposes Python bindings.
 Functions:
 encode(x) -> y:
     x: 1D int32 tensor.
-    y: 1D uint8 tensor.
+    y: 1D uint32 tensor.
 decode(y) -> x:
-    y: 1D uint8 tensor.
+    y: 1D uint32 tensor.
     x: 1D int32 tensor.
 """
